Split geoparquet and bbf into dedicated file-format crates
Extract the GeoParquet write integration into beacon-arrow-geoparquet and the BBF DataFusion integration into beacon-arrow-bbf, following the established beacon-arrow-* crate pattern (types under a `datafusion` submodule).

beacon-formats keeps thin re-export shims for geo_parquet and bbf so existing consumers compile unchanged. The shared rlimit helpers (max_open_fd/file_open_parallelism) move to beacon-common::file_descriptors so the new crates and the remaining arrow/csv/parquet wrappers can share them.
