+++ beacon-common/Cargo.toml
@@ -22,3 +22,4 @@ object_store = { workspace = true }
 futures = { workspace=true}
 futures-util = { workspace=true}
 tracing = { workspace = true }
+rlimit = "0.11.0"
