+++ beacon-common/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod cf_time;
 pub mod error;
+pub mod file_descriptors;
 pub mod listing_url;
 pub mod schema_table_provider;
 pub mod super_table;
