+++ beacon-file-formats/beacon-arrow-bbf/src/datafusion/metrics.rs
