+++ beacon-file-formats/beacon-arrow-bbf/src/datafusion/stream_share.rs
