+++ beacon-file-formats/beacon-formats/Cargo.toml
@@ -7,9 +7,6 @@ edition = "2024"
 arrow = { workspace = true }
 object_store = { workspace = true }
 datafusion = { workspace = true }
-geoarrow = { workspace = true }
-geoarrow-array = { workspace = true }
-geoparquet = { workspace = true }
 parquet = { workspace = true }
 ordered-float = { workspace = true }
 
@@ -38,7 +35,6 @@ which = "8.0.0"
 num-traits = "0.2.19"
 ndarray = { version = "0.17" }
 once_cell = "1.21.3"
-rlimit = "0.11.0"
 chrono = { workspace = true }
 
 beacon-common = { path = "../../beacon-common" }
@@ -47,13 +43,11 @@ beacon-arrow-netcdf = { path = "../beacon-arrow-netcdf" }
 beacon-arrow-odv = { path = "../beacon-arrow-odv" }
 beacon-arrow-tiff = { path = "../beacon-arrow-tiff" }
 beacon-arrow-zarr = { path = "../beacon-arrow-zarr" }
+beacon-arrow-geoparquet = { path = "../beacon-arrow-geoparquet" }
+beacon-arrow-bbf = { path = "../beacon-arrow-bbf" }
 beacon-config = { path = "../../beacon-config" }
 beacon-object-storage = { path = "../../beacon-object-storage" }
 beacon-datafusion-ext = { path = "../../beacon-datafusion-ext" }
-nd-arrow-array = { git = "https://github.com/maris-development/nd-arrow-array.git", branch = "main", version = "2.0.0", default-features = false, features = ["arrow-58"] }
-
-# Add beacon binary format
-beacon-binary-format = { path = "../beacon-binary-format", default-features = false, features = ["arrow-58", "object-store-13"] }
 arrow-flight = { workspace = true }
 tonic = "0.14"
 prost = "0.14"
